+++ backends/xla/builder.go
@@ -305,16 +305,16 @@ func convertPadAxis(pad backends.PadAxis) (xlaPad xlabuilder.PadAxis) {
 	return
 }
 
-func convertFFTType(fftType backends.FFTType) xla_data.FftType {
+func convertFFTType(fftType backends.FFTType) xlabuilder.FFTType {
 	switch fftType {
 	case backends.FFTForward:
-		return xla_data.FftType_FFT
+		return xlabuilder.FFTType_FFT
 	case backends.FFTInverse:
-		return xla_data.FftType_IFFT
+		return xlabuilder.FFTType_IFFT
 	case backends.FFTForwardReal:
-		return xla_data.FftType_RFFT
+		return xlabuilder.FFTType_RFFT
 	case backends.FFTInverseReal:
-		return xla_data.FftType_IRFFT
+		return xlabuilder.FFTType_IRFFT
 	default:
 		exceptions.Panicf("fft type %s is not supported", fftType)
 		panic(nil) // To quiet IDE warning.