+++ docs/CHANGELOG.md
@@ -1,5 +1,10 @@
 # GoMLX changelog
 
+# Next
+
+* Package `xla`
+  * Remove dependencies to `gopjrt` internal protos: requires updated `Gopjrt`.
+
 # v0.19.3: 2024/05/20 Many SimpleGo improvements. 
 
 * v0.19.2 skipped ... issues with the release.