+++ docs/CHANGELOG.md
@@ -2,11 +2,11 @@
 
 # v0.19.4: 2024/05/24 added Vector Neural Networks (VNNs)
 
+* Vector Neural Networks (VNN): allows one to build 3D rotation (SO(3)) equivariant and/or invariant networks. See package `ml/layers/vnn`.
 * Package `xla`
   * Remove dependencies to `gopjrt` internal protos: requires updated `Gopjrt`.
 * Package `tensors`
   * Fixed pretty-print of booleans.
-* Vector Neural Networks (VNN): allows one to build 3D rotation (SO(3)) equivariant and/or invariant networks. See package `ml/layers/vnn`. 
 
 # v0.19.3: 2024/05/20 Many SimpleGo improvements.